checkpoint - working codegen

+++ include/ast_io.h
@@ -8,14 +8,14 @@
 #define AST_MAGIC "ZAST"
 #define AST_FORMAT_VERSION 1
 
-int ast_write_u8(output_t out, uint8_t value);
-int ast_write_u16(output_t out, uint16_t value);
-int ast_write_u32(output_t out, uint32_t value);
-int ast_write_i16(output_t out, int16_t value);
+int8_t ast_write_u8(output_t out, uint8_t value);
+int8_t ast_write_u16(output_t out, uint16_t value);
+int8_t ast_write_u32(output_t out, uint32_t value);
+int8_t ast_write_i16(output_t out, int16_t value);
 
-int ast_read_u8(reader_t* reader, uint8_t* out);
-int ast_read_u16(reader_t* reader, uint16_t* out);
-int ast_read_u32(reader_t* reader, uint32_t* out);
-int ast_read_i16(reader_t* reader, int16_t* out);
+int8_t ast_read_u8(reader_t* reader, uint8_t* out);
+int8_t ast_read_u16(reader_t* reader, uint16_t* out);
+int8_t ast_read_u32(reader_t* reader, uint32_t* out);
+int8_t ast_read_i16(reader_t* reader, int16_t* out);
 
 #endif /* AST_IO_H */
